fix xformers attn_bias dtype mismatch under autocast
The pre-expanded (B, H, Lq, Lk) local mask was passed to xformers as-is
without casting to query.dtype. Under bfloat16 autocast the queries are
bf16 while the stored mask may be float32, triggering:
  RuntimeError: invalid dtype for bias - should match query's dtype

Cast attn_bias to query.dtype at the point of use, consistent with the
other two branches that already call .to(query.dtype).

+++ src/cehrgpt/models/gpt2.py
@@ -328,8 +328,8 @@ class GPT2FlashAttention(GPT2Attention):
                 # across all layers).  If so, use it directly.  Otherwise expand here
                 # — this covers the decode path (B, 1, 1, KV_len) and cross-attention.
                 if attention_mask.shape[1] == self.num_heads:
-                    # Already pre-expanded and aligned — use as-is.
-                    attn_bias = attention_mask
+                    # Already pre-expanded and aligned — cast to query dtype to satisfy xformers.
+                    attn_bias = attention_mask.to(query.dtype)
                 else:
                     # (B, 1, Lq, Lk) — expand with xformers cutlass alignment.
                     # stride(-2) of the underlying allocation must be a multiple of 8.
